[GENERAL] - Prep work.
Getting the 32-bit build up and running again, complete with process code and able to run CLI etc.
Aiming to get stdlib, and userspace programs (CLI, BASIC, HELLO) all building in 32/64-bit.

Work to do yet, but checking in the current state as it is deploying successfully in full in 32-bit at this point.