[32-BIT USPI] - Reintroduced
USPI has been reintroduced at this point into the 32-bit build successfully, so the Pi
fires up, launches CLI, and the keyboard can be used to enter "HELLO" to then invoke the
hello binary. All successful.

Next step will be to figure out how to get USPI working in the 64-bit equivalent, either through
updating the existing USPI, or grabbing the latest USPI which is meant to be 64-bit compatible and
trying that.