[64-BIT] - FULLY WORKING
The 64-bit build is done. It's running, it runs processes, the systimer is ticking away,
there are no crashes because EL1 stack is kept under control.

At this point what can be seen is the 32-bit build is now breaking. It looks like a keyboard
(USPi) related issue. But this now needs to be investigated under its own merits, looking to see
what has changed recently in getting to the 64-bit build working, that has perhaps had an adverse
effect on the 32-bit version.