[32/64-BIT] Both working fine now.
There was _kernel_systimer_init being called in the 32-bit version. The _kernel_process_start()
was not returning correctly, and as _kernel_systimer_init() was immediately after _kernel_process_start()
in memory, the end result was the systimer firing up, basically.

Now fixed - 32 and 64-bit builds working fine and consistent. Work now needs to start on cleaning up the
code base, getting USPi flexibly buiding as 32/64-bit, and eradicating a minor handful of compiler warnings
on the 64-bit USPi build (preferable to have a clean, stress-free build for people).