[32/64-BIT & PI 3B/3B+] All builds tested.
A minor change to the USPi code as there was a continuing misalignment issue with part of a data structure
and the simplest way to circumvent it at this stage is to _kernel_memcpy the offending part.

systimer no longer printing out "." all the time either.

This build has been tested and found to build, deploy and behave consistently on the following platforms:

* Pi 3B 32-bit
* Pi 3B 64-bit
* Pi 3B+ 32-bit
* Pi 3B+ 64-bit

This is a baseline release effectively so the version number will now be bumped up significantly to 0.5.